update display for NLP task
